Update Magic SDK extensions typing to support PNP
