+++ packages/@magic-sdk/pnp/src/types.ts
@@ -1,0 +1,12 @@
+import type { Magic } from 'magic-sdk/dist/types/index.es';
+import type OAuthExtension from '@magic-ext/oauth/dist/es/index.cdn';
+import type CDNMagic from 'magic-sdk/dist/types/index.cdn';
+
+declare global {
+  interface Window {
+    Magic: typeof CDNMagic;
+    MagicOAuthExtension?: typeof OAuthExtension;
+  }
+}
+
+export type { Magic, OAuthExtension };
