+++ packages/@magic-sdk/provider/src/index.ts
@@ -1,8 +1,8 @@
 /* istanbul ignore file */
 
-export { SDKBase, MagicSDKAdditionalConfiguration } from './core/sdk';
+export { SDKBase, MagicSDKAdditionalConfiguration, MagicSDKExtensionsOption } from './core/sdk';
 export { createSDK } from './core/sdk-environment';
 export { ViewController } from './core/view-controller';
 export * from './core/sdk-exceptions';
-export { Extension } from './modules/base-extension';
+export { Extension, WithExtensions, InstanceWithExtensions } from './modules/base-extension';
 export * from './util';
